creating separate component for technical experiences
