+++ src/components/TechnicalExperiences/index.js
@@ -1,0 +1,2 @@
+import TechnicalExperiences from './TechnicalExperiences';
+export default TechnicalExperiences;
